Auth0 fully integrated.
When you visit the homepage, the auth0 middleware checks if there's a code query parameter.

If there isn't, redirects the user to the login endpoint, which will create an authorization URL according to the auth0 documentation regarding user authorization.
The user will then be redirected to the homepage with a code attached as a query.

If there is a query attached to the homepage url, the middleware will try to trade the code for an access token. If successful, it will connect to the server API and try to access a secured resource.

The API receives the token, uses jwtCheck to verify its authenticity, and responds with the secured data.

The app then responds to the user with a json of the data provided by th api.

Removed permissions because theyre unnecessary. All users can read, modify and delete data.