Updating character get endpoint and now can switch between three results.
The request query will be filtered to fields subject to queries, established by var fields, then:

Option a)
If the request query has an ID field, it will ignore all other field and try to resolve the query using that field only.
- If the ID is not a number will respond with bad request.
- If the database doesn't find the ID, will respond with 404.
- If the database finds the ID, will respond with the resource.

Option b) If req.query has any other kind of queries, will use all of them to return a filtered database. In this case, an empty array wont constitute a 404.

Option c) In any other case, the API will respond with the Character list.